Manage Errors (#12)
# Motivation

Improve how the library manages errors and exposes them to users.

The main improvement is to differentiate technical erros, which will
output at onError, from not getting a credential for other reasons,
which go to onSuccess with a different outout.

# Changes

* Change onSuccess param type.
* Check whether user closes the window without finishing flow. We
consider this as a technical error and ends up in `onError`.
* Changed `currentFlows` from a `Set` to a `Map` with the status of the
current flow. Then the status is changed and checked during the flow.

# Tests

* Remove skipped test `"calls onError with timeout error if flow doesn't
start in five seconds"`. AuthClient doesn't manage it; therefore, I
think that it makes sense we don't do it either.
* Implement test `"calls onError if user closes identity provider
window"`.
* Add a test `"should not call onError when window closes after
successful flow"`.
* Change the test `"calls onError when the credential fails"` to check
that it calls `onSuccess`.