Add GitHub actions workflow to create a release when pushing a release tag (#39)
This PR sets up a workflow to publish a release when a tag with a
`release-` prefix is pushed
to the repository.

The created release will have the dummy issuer and dummy relying party
attached as artifacts. These will later be referenced for `dfx deps`
feature.

The release will only be created as a draft to allow for manual
modifications before actually publishing.