build: support for webpack and editors (#56)
# Motivation

It was reported that the library does not work with Next.JS but, we were
not abole to reproduce the issue. However, the developer is using a
custom Webpack, that's why we can actually assume that the issue related
to this bundle.

On another side, the issue does not seems to reside solely in the build
but, also when imported in an editor as Webstorm or VS Code does not
seem able to understand the lib is modular.

That's why, there are various things we should do to fix the problem.

# Webpack

Webpack seems to require a `sideEffects` flag in `package.json` to
enable tree shaking for a library
([documentation](https://webpack.js.org/guides/tree-shaking/)). This
was, for example, introduced in
[three.js](https://github.com/mrdoob/three.js/issues/16059).

I'm also unsure whether Webpack requires a `main` field. This PR adds
the main entry, even though it will still resolve to an index that
throws an error, since we are providing a modular library.

# Import and validation

When using `exports` in `package.json` it seems that editor have troube
understanding that the modules are not shipped at the root of the
library. I did not find a solution therefore decided to adapt the way we
provide the information by moving `package.json`, and other metadata
files` to the `dist` folder. That way the package can reference files at
the same level.

⚠️ This comes with an important change. The library after `npm run
build` should not be published anymore from the root of the library but,
from the `dist` folder. That is why the pipelines had to be updated.

# Other changes

While fixing those issues I noticed few other minor problems. The
`mocks.ts` was packaged in the lib and the file `LICENSE` was missing.