build: add jsdom as dev dependency (#90)
# Motivation

CI `checks` are failing - e.g.
https://github.com/dfinity/verifiable-credentials-sdk/actions/runs/16936244767/job/47993367472?pr=87.
It looks like vite now explicitely requires jsdom to be specified. Weird
as it seems we did not upgrade vitest recently. Nevertheless, this PR is
meant to unlock the CI.

# Notes

It seems that pinning Node LTS also resolves it - see
https://github.com/dfinity/verifiable-credentials-sdk/pull/89 - but
still, given that vitest anyway requires jsdom, not too bad to list it
anyway.

# Changes

-`npm i jsdom -D`